refactor build 0.01
First push to the new "refactor" branch, with the aim to refactor code from a single "main.c" to several header and source files.
This version has linker errors.
