+++ .gitignore
@@ -1,0 +1,2 @@
+
+data/~$Weather_Data.xlsx
