+++ src/enemy.h
@@ -1,0 +1,6 @@
+#ifndef enemy_h
+#define enemy_h
+
+
+
+#endif
